requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
pydantic
scikit-learn
httpx
python-dotenv

pydantic
pydantic-settings

pymongo[srv]
passlib[bcrypt]

python-jose[cryptography]